Print out the fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
 
 int main(void)
 {
-	int num1;
-	int num2;
-	int sum;
+	long num1;
+	long num2;
+	long sum;
 	int i;
 
 	num1 = 1;
